+++ lib/index.js
@@ -1,4 +1,6 @@
-var libvirt_native = require('bindings')('libvirt');
+let libvirt_native = require('bindings')('libvirt');
+let virConnect = require('../generated/connect');
+let virDomain = require('../generated/domain');
 
 /**
  * This function should be called first to get a connection to the
@@ -27,44 +29,6 @@ function open(name) {
     let ret = libvirt_native.virConnectOpen(name);
     if(!ret) throw new Error('virConnectOpen() failed');
     return new virConnect(ret);
-}
-
-class virConnect {
-    constructor(conn) {
-        this.conn = conn;
-    }
-
-    /**
-     * Try to lookup a domain on the given hypervisor based on its name.
-     *
-     * virDomainFree should be used to free the resources after the
-     * domain object is no longer needed.
-     *
-     * @param {String} name
-     * @returns {Object} virDomain
-     */
-    lookupByName(name) {
-        let ret = libvirt_native.virDomainLookupByName(this.conn, name);
-        if(!ret) throw new Error('virDomainLookupByName() failed');
-        return new virDomain(this, ret);
-    }
-}
-
-class virDomain {
-    constructor(conn, domain) {
-        this.conn = conn;
-        this.domain = domain;
-    }
-
-    /**
-     * Get the hypervisor ID number for the domain
-     *
-     * @return {Integer}
-     */
-    getId() {
-        let ret = libvirt_native.virDomainGetID(this.domain);
-        return ret;
-    }
 }
